refactor: altera testes de exclusão
