+++ src/routes/tests/sebo.routes.test.ts
@@ -66,4 +66,14 @@ describe('Sebo Routes', () => {
     expect(response.body.message).toBe('mensagem teste para atualização');
     expect(seboController.update).toHaveBeenCalledTimes(1);
   });
+
+  it('deletes a sebo by id', async () => {
+    (seboController.delete as jest.Mock).mockImplementationOnce((req, res) => {
+      res.status(200).send({ message: 'mensagem teste para exclusão' });
+    });
+
+    const response = await request(app).delete('/sebos/1');
+    expect(response.status).toBe(200);
+    expect(response.body).toEqual({ message: 'mensagem teste para exclusão' });
+  });
 });
